fix CRM-6035

+++ CRM/Activity/BAO/Activity.php
@@ -364,8 +364,10 @@ class CRM_Activity_BAO_Activity extends CRM_Activity_DAO_Activity
             $assignmentParams = array( 'activity_id'         => $activityId );
 
             if ( is_array( $params['assignee_contact_id'] ) ) {
-                // first delete existing assignments if any
-                self::deleteActivityAssignment( $activityId );
+                if ( CRM_Utils_Array::value( 'deleteActivityAssignment', $params, true ) ) {
+                    // first delete existing assignments if any
+                    self::deleteActivityAssignment( $activityId );
+                }
 
                 foreach ( $params['assignee_contact_id'] as $acID ) {
                     if ( $acID ) {
@@ -394,7 +396,9 @@ class CRM_Activity_BAO_Activity extends CRM_Activity_DAO_Activity
                 }
             }
         } else {       
-            self::deleteActivityAssignment( $activityId );
+            if ( CRM_Utils_Array::value( 'deleteActivityAssignment', $params, true ) ) {
+                self::deleteActivityAssignment( $activityId );
+            }
         }
 
         if( is_a( $resultAssignment, 'CRM_Core_Error' ) ) {
@@ -409,8 +413,10 @@ class CRM_Activity_BAO_Activity extends CRM_Activity_DAO_Activity
             $targetParams = array( 'activity_id'       => $activityId );
             $resultTarget = array( );
             if ( is_array( $params['target_contact_id'] ) ) {
-                // first delete existing targets if any
-                self::deleteActivityTarget( $activityId );
+                if ( CRM_Utils_Array::value( 'deleteActivityTarget', $params, true ) ) {
+                    // first delete existing targets if any
+                    self::deleteActivityTarget( $activityId );
+                }
 
                 foreach ( $params['target_contact_id'] as $tid ) {
                     if ( $tid ) {
@@ -439,7 +445,9 @@ class CRM_Activity_BAO_Activity extends CRM_Activity_DAO_Activity
                 }
             }
         } else {
-            self::deleteActivityTarget( $activityId );
+            if ( CRM_Utils_Array::value( 'deleteActivityTarget', $params, true ) ) {
+                self::deleteActivityTarget( $activityId );
+            }
         }
 
         // write to changelog before transation is committed/rolled
